+++ crates/crb-runtime/src/lib.rs
@@ -4,7 +4,7 @@ pub mod interruptor;
 pub mod runtime;
 pub mod task;
 
-pub use context::{Context, ManagedContext};
+pub use context::{ManagedContext, ReachableContext};
 pub use error::Failures;
 pub use interruptor::{Controller, Interruptor, RegistrationTaken};
 pub use runtime::{InteractiveRuntime, Runtime};